銃弾当たり判定修正

+++ 卒業制作作品/卒業制作作品/ObjRocketLauncherAttack.cpp
@@ -60,6 +60,9 @@ void CObjRocketLauncherAttack::Init()
 	//爆発・血しぶき用描画サイズ
 	m_exp_blood_dst_size = 320.0f;
 
+	//HitBox削除フラグ
+	m_HitBox_Delete = false;
+
 	if (m_RLr == 0 || m_RLr == 180)
 	{
 		//当たり判定用HitBoxを作成
@@ -71,7 +74,6 @@ void CObjRocketLauncherAttack::Init()
 		Hits::SetHitBox(this, m_RLx, m_RLy, 70, 32, ELEMENT_RED, OBJ_ROCKETLAUNCHERATTACK, 2);
 	}
 	
-
 }
 
 //アクション
@@ -128,8 +130,8 @@ void CObjRocketLauncherAttack::Action()
 			CObjExplosion* obj_bs = new CObjExplosion(m_RLx - 128, m_RLy - 128, m_exp_blood_dst_size, ((UserData*)Save::GetData())->RL_Attack);
 			Objs::InsertObj(obj_bs, OBJ_EXPLOSION, 9);
 			Audio::Start(9);
-			this->SetStatus(false); //オブジェクト破棄
-			Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+
+			m_HitBox_Delete = true;
 		}
 		else if (m_RLx > hx + 32 + 64 * m_Distance_max)
 		{
@@ -137,8 +139,8 @@ void CObjRocketLauncherAttack::Action()
 			CObjExplosion* obj_bs = new CObjExplosion(m_RLx - 128, m_RLy - 128, m_exp_blood_dst_size, ((UserData*)Save::GetData())->RL_Attack);
 			Objs::InsertObj(obj_bs, OBJ_EXPLOSION, 9);
 			Audio::Start(9);
-			this->SetStatus(false); //オブジェクト破棄
-			Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+
+			m_HitBox_Delete = true;
 		}
 		if (m_RLy < hy - 64 * m_Distance_max)
 		{
@@ -146,8 +148,8 @@ void CObjRocketLauncherAttack::Action()
 			CObjExplosion* obj_bs = new CObjExplosion(m_RLx - 128, m_RLy - 128, m_exp_blood_dst_size, ((UserData*)Save::GetData())->RL_Attack);
 			Objs::InsertObj(obj_bs, OBJ_EXPLOSION, 9);
 			Audio::Start(9);
-			this->SetStatus(false); //オブジェクト破棄
-			Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+
+			m_HitBox_Delete = true;
 		}
 		else if (m_RLy > hy + 32 + 64 * m_Distance_max)
 		{
@@ -155,8 +157,8 @@ void CObjRocketLauncherAttack::Action()
 			CObjExplosion* obj_bs = new CObjExplosion(m_RLx - 128, m_RLy - 128, m_exp_blood_dst_size, ((UserData*)Save::GetData())->RL_Attack);
 			Objs::InsertObj(obj_bs, OBJ_EXPLOSION, 9);
 			Audio::Start(9);
-			this->SetStatus(false); //オブジェクト破棄
-			Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+
+			m_HitBox_Delete = true;
 		}
 	}
 
@@ -175,8 +177,8 @@ void CObjRocketLauncherAttack::Action()
 			CObjExplosion* obj_bs = new CObjExplosion(m_RLx - 140, m_RLy - 140, m_exp_blood_dst_size, ((UserData*)Save::GetData())->RL_Attack);
 			Objs::InsertObj(obj_bs, OBJ_EXPLOSION, 9);
 			Audio::Start(9);
-			this->SetStatus(false); //オブジェクト破棄
-			Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+
+			m_HitBox_Delete = true;
 		}
 	}	
 	//壁オブジェクトと接触するとオブジェクト破棄
@@ -188,8 +190,8 @@ void CObjRocketLauncherAttack::Action()
 		CObjExplosion* obj_bs = new CObjExplosion(m_RLx - 140, m_RLy - 140, m_exp_blood_dst_size, ((UserData*)Save::GetData())->RL_Attack);
 		Objs::InsertObj(obj_bs, OBJ_EXPLOSION, 9);
 		Audio::Start(9);
-		this->SetStatus(false); //オブジェクト破棄
-		Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+
+		m_HitBox_Delete = true;
 	}
 	//フィールドエレメントと接触すると削除
 	if (hit_rl->CheckElementHit(ELEMENT_FIELD) == true || hit_rl->CheckElementHit(ELEMENT_FIELD2) == true)
@@ -208,9 +210,19 @@ void CObjRocketLauncherAttack::Action()
 			CObjExplosion* obj_bs = new CObjExplosion(m_RLx - 140, m_RLy - 140, m_exp_blood_dst_size, ((UserData*)Save::GetData())->RL_Attack);
 			Objs::InsertObj(obj_bs, OBJ_EXPLOSION, 9);
 			Audio::Start(9);
-			this->SetStatus(false); //オブジェクト破棄
-			Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+
+			m_HitBox_Delete = true;
 		}
+	}
+
+
+	//削除処理
+	if (m_HitBox_Delete == true)
+	{
+		this->SetStatus(false); //オブジェクト破棄
+		Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+
+		m_HitBox_Delete = false; //初期化
 	}
 }
